This fix corrects a number of issues with the plugin, as forked from Jerome O'Flaherty's git repo in mid-May 2023 (git commit 1477ebc from 29-Nov-2022.)
This change replaces the 'net-ping' package with the 'ping' package.  The 'net-ping' package is functionally
superior, but it depends on the 'raw-socket' package, and that suffers from raw-socket issue #70, which causes the
node thread running the ping to suddenly exit. Speculation is that this is caused by garbage collection
in nodejs itself.  This issue remains unresolved for over a year.

The exit of the thread causes the plugin server discovery to fail, causing the any attempt to used Plex
functionality to default to PlexAmp running on 127.0.0.1 (localhost.)  This produces a "server not found"
error unless PlexAmp is locally installed.

The plex.tv URL used to discover server information is also changed from https://plex.tv/pms/servers to
https://plex.tv/pms/resources.   The former link only shows servers that have Remote Access enabled.  Since
many local Plex servers do not have that enabled, it causes server discovery to fail.  Again, this results
in a "server not found" error when no local servers that respond to ICMP ping are found.

Discovered server records are now filtered to a standard JSON structure interface in the plexcloud getServers(),
routine rather than directly exposing the raw output of XML converted to JSON.  The currectly recognized fields are:

{
   "name": "",               # Server name, according to plex.tv (not DNS name)
   "protocol": "",           # Access protocol -- http or https
   "address": "",            # Server IP address
   "port": "",               # Server TCP port number
   "local": ""               # Boolean indicating that the address is local to the host where Plex runs when true,
                             #   or is a Remote Access IP address if false
}

The tests for this plugin used the library name "Ballyoran" to test Plex access.  This is changed to the environment
variable name LIBRARYNAME.

There are also a number of minor code fixes that just prevented the code from fully functioning.

This should allow the plugin to more fully function in line with Jerome O'Flaherty's original intentions.

The plugin setup still chooses the first server it finds.  Allowing users to choose between multiple Plex servers
and how to verify which servers are actually running are the next big issues to be resolved in this code, I think.