Better album support
Added a list of all albums to the top level Plex screen

In the process of testing the new Albums list, I discovered that the transitions to show album contents were failing.
The failures occurred when looking up a list of related albums to the selected album.  That caused an access
to an undefined property for album information returned by my local Plex server.  The search for related albums
now continues only if the necessary property exists.