Server detection update
While using ICMP ping to detect whether a media server is reachable is a logical choice,
having ICMP ping enabled for a server on the Internet is not absolutely necessary. Cloud
services like Amazon EC2 do not enable ICMP by default for public addresses to avoid
DDoS attacks.  Having ICMP ping enabled for local media servers on private servers is a
safe assumption, but not necessarily for remote access servers.

Changed the ICMP ping to a http header request against the plex http port on the media
server.  This _must_ work to contact the media server.  The request has a 1 second timeout.
The coding is it a little weird because of how Plex works.

The ping is an unauthenticated http request, so a successful request will come back as 401
Unauthorized.  The 'request-promise' library treats 401 as an error, so the error exception
generated by the request must be handled in a 'catch' section.  On the off chance that the
http request comes back as a success in future, any 2XX response is treated as a success.

Beyond that, code in plexpinauth.js is modified to use kew like much of the rest of the code.
This allows for substantial cleanup of the code in getPinAndUpdateConfig in index.js.

Much of the code in getPinAndUpdateConfig and getUIConfig is cleaned up.  It lowers
the nesting levels of both by making better use of promises (.then and .fail) and by
replacing long references to sections in UIConfig.json with useful names.

testPlex.js is modified to test the changes in structure to getPinAndUpdateConfig and
use the http ping.

One problem I discovered after I got multiple media servers and libraries working is that the label for the
Music Library showing what the current saved setting did not work.  This turned out to be because the filter
compares both the content and the type of saved and possible properties of the music library:

var selectedLibrary = libraries.filter((filteredLibrary) => filteredLibrary.hostname === hostname &&
                      filteredLibrary.port === port && filteredLibrary.libraryTitle === libraryName);

The problem turned out to be that port was a number type (despite being loaded from a string) and
filteredLibrary.port was a string type.   The fix: make sure that port is a string type.